working enrollment,manager,test

+++ app/src/main/java/model/user/Professor.java
@@ -2,7 +2,7 @@ package model.user;
 
 public class Professor extends User{
 
-    public Professor(int id, String name, String password) {
+    public Professor(long id, String name, String password) {
         super(id, name, password, 0);
     }
